Add a game controller to play the game via console
Player object added.

The game controller alternates between each
player's turn time automatically. It also
shows which player is supposed to play at a given
moment.

The method "insertToken" has now a return value.
It's been added because the controller uses it
to validate if a certain play from the user is
valid or not.