Add feature to check winner or possible draw
It is not desirable to have direct access on
_board from outside "Gameboard" function, which
is the reason why "getTokenAt" method is
now present.

GameController function returns added isGameOver
attribute to the returned object so that so that
an external utility is able to tell if game has
finished yet.

If the game is finished, nothing else happens.
Line 165 causes this behavior.